Splits rules into modules
* uses object assign polyfill for settings
* moves declarations and at rules into module

Closes #156
